requirements.txt
streamlit==1.24.1
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
